remap Y y$
The commands yy, dd, and cc operate on the whole line. The commands C and D change/delete from the cursor to EOL, but Y yanks the whole line instead. This mapping will align Y with the behavior of C and D.

This is actually a recommeded remapping. See :help Y